chore: split forms.ts and forms.server.ts
