+++ sveltekit/src/routes/account/add/+page.server.ts
@@ -1,4 +1,5 @@
-import { getSelectAccountTypes, selectBalanceGroups } from '$lib/helpers/forms';
+import { getSelectAccountTypes } from '$lib/helpers/forms.server';
+import { selectBalanceGroups } from '$lib/helpers/forms';
 
 export const load = async () => {
 	const selectAccountTypes = await getSelectAccountTypes();
