+++ sveltekit/src/routes/asset/add/+page.server.ts
@@ -1,8 +1,5 @@
-import {
-	getSelectAssetTypes,
-	getQuantifiableAssetTypes,
-	selectBalanceGroups
-} from '$lib/helpers/forms';
+import { getSelectAssetTypes, getQuantifiableAssetTypes } from '$lib/helpers/forms.server';
+import { selectBalanceGroups } from '$lib/helpers/forms';
 
 export const load = async () => {
 	const selectAssetTypes = await getSelectAssetTypes();
